Driver limits moved to base driver. Cleaner code.
